+++ src/graphene_elastic/arrayconnection.py
@@ -12,9 +12,8 @@ from .relay.connectiontypes import Connection
 
 from .logging import logger
 
-__title__ = 'graphene_elastic.arrayconnection'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2019-2020 Artur Barseghyan'
+__copyright__ = '2019-2022 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'connection_from_list_slice',