+++ src/graphene_elastic/compat.py
@@ -4,9 +4,8 @@
 # module is not supposed to solve all transition issues for you. Better move to
 # Elastic 7.x as soon as possible.
 # """
-# __title__ = 'graphene_elastic.compat'
 # __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-# __copyright__ = '2019-2020 Artur Barseghyan'
+# __copyright__ = '2019-2022 Artur Barseghyan'
 # __license__ = 'GPL 2.0/LGPL 2.1'
 # __all__ = (
 #     'nested_sort_entry',