+++ src/graphene_elastic/enums.py
@@ -1,11 +1,13 @@
 from enum import Enum
 import graphene
 
-__title__ = "graphene_elastic.enums"
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2019-2020 Artur Barseghyan"
+__copyright__ = "2019-2022 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
-__all__ = ("NoValue", "convert_list_to_enum")
+__all__ = (
+    "NoValue",
+    "convert_list_to_enum",
+)
 
 
 class NoValue(Enum):