+++ src/graphene_elastic/filter_backends/__init__.py
@@ -7,7 +7,6 @@ from .score import *
 from .search import *
 from .source import *
 
-__title__ = 'graphene_elastic.filter_backends'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2019-2020 Artur Barseghyan'
+__copyright__ = '2019-2022 Artur Barseghyan'
 __license__ = 'GPL-2.0-only OR LGPL-2.1-or-later'