+++ src/graphene_elastic/filter_backends/faceted_search/common.py
@@ -2,8 +2,8 @@ from copy import deepcopy
 # from collections import OrderedDict
 import enum
 
-from elasticsearch_dsl import TermsFacet
-from elasticsearch_dsl.query import Q
+from anysearch.search_dsl import TermsFacet
+from anysearch.search_dsl.query import Q
 
 import graphene
 from stringcase import pascalcase as to_pascal_case
@@ -12,9 +12,8 @@ from stringcase import pascalcase as to_pascal_case
 from ...constants import DYNAMIC_CLASS_NAME_PREFIX
 from ..base import BaseBackend
 
-__title__ = 'graphene_elastic.filter_backends.faceted_search.common'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2019-2020 Artur Barseghyan'
+__copyright__ = '2019-2022 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'FacetedSearchFilterBackend',