+++ src/graphene_elastic/filter_backends/filtering/common.py
@@ -27,9 +27,8 @@ from ...constants import (
 from .mixins import FilteringFilterMixin
 from .queries import LOOKUP_FILTER_MAPPING
 
-__title__ = "graphene_elastic.filter_backends.filtering.common"
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
-__copyright__ = "2019-2020 Artur Barseghyan"
+__copyright__ = "2019-2022 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = ("FilteringFilterBackend",)