+++ src/graphene_elastic/filter_backends/search/query_backends/base.py
@@ -1,6 +1,5 @@
 import six
 
-__title__ = 'graphene_elastic.filter_backends.search.query_backends.base'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2019 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'