+++ src/graphene_elastic/filter_backends/search/query_backends/simple_query_string.py
@@ -1,13 +1,11 @@
 # import copy
 #
-# from elasticsearch_dsl.query import Q
+# from anysearch.search_dsl.query import Q
 #
 # from .base import BaseSearchQueryBackend
 #
-# __title__ = 'graphene_elastic.filter_backends.search.' \
-#             'query_backends.simple_query_string'
 # __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-# __copyright__ = '2019 Artur Barseghyan'
+# __copyright__ = '2019-2022 Artur Barseghyan'
 # __license__ = 'GPL 2.0/LGPL 2.1'
 # __all__ = ('SimpleQueryStringQueryBackend',)
 #