+++ src/graphene_elastic/versions.py
@@ -5,9 +5,8 @@ including (LTE and GTE).
 
 from distutils.version import LooseVersion
 
-__title__ = 'graphene_elastic.versions'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2019-2020 Artur Barseghyan'
+__copyright__ = '2019-2022 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = [
     'get_elasticsearch_version',