+++ CHANGELOG.rst
@@ -15,6 +15,25 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.8
+---
+2022-07-31
+
+.. note::
+
+    The ``elasticsearch`` and ``elasticsearch-dsl`` packages are no longer
+    installed by default. You must either install them explicitly in your
+    requirements or install alongside this package as optional dependencies as
+    follows: ```pip install graphene-elastic[elasticsearch]```.
+    Alternatively, you can use ``opensearch-py`` and ``opensearch-dsl``.
+    You would then need to install the ``opensearch-py`` and ``opensearch-dsl``
+    packages explicitly in your requirements or install alongside this
+    package as optional dependencies as
+    follows: ```pip install graphene-elastic[opensearch]```.
+
+- Added support for OpenSearch (1.x and 2.x).
+- Belated migration to GitHub Actions.
+
 0.7
 ---
 2021-03-08